+++ Reqnroll/Infrastructure/TestExecutionEngine.cs
@@ -192,9 +192,9 @@ namespace Reqnroll.Infrastructure
             }
             catch (Exception)
             {
-                // Removed code that sets the ScenarioContext.TestError and ScenarioContext.ScenarioExecutionStatus, because
-                // FireEventsAsync called by FireScenarioEventsAsync sets *Context.TestError and ScenarioContext.ScenarioExecutionStatus.
-                // The fact that the exception is not rethrown here is suspicious, but it is the current behavior. We need to check it eventually.
+                // When StopAtFirstError is false (default), we do not rethrow the exception, because it will be handled in OnAfterLastStepAsync
+                if (_reqnrollConfiguration.StopAtFirstError)
+                    throw;
             }
         }