+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ## Bug fixes:
 
 * Fix: Exceptions from before scenario plugin hook events (RuntimePluginTestExecutionLifecycleEvents.BeforeScenario) are not shown (#856)
+* Fix: Before scenario hooks are not raised immediately when `runtime/stopAtFirstError` is true, but handled delayed (#857)
 
 *Contributors of this release (in alphabetical order):* @chekkan, @Code-Grump, @gasparnagy, @konarx